Use Py_SET_TYPE
In Python 3.11a0dev (main branch), Py_TYPE is a function instead
of a macro, which means it cannot be assigned to.
For assignments, Py_SET_TYPE should be used. That was only added
in Python 3.9, but for older versions we can use a simple macro.

It's possible that this change will be reverted before Python 3.11
is released, but still, assigning to Py_TYPE is deprecated.