Fix error handling in MGLError_SetTrace
When MGLError_SetTrace was called with an active exception
(such as from MGLFramebuffer_set_scissor after passing a
non-int element), PyImport_ImportModule was called with an
exception set.
The effect on the import machinery is unpredictable: generally
the first check of PyErr_Occurred (rather than of the C return
value) will assume a function failed.
In Python 3.10 this leads to a fatal crash.

Clear the old exception *first* since we replace it with a new one.
(The super-proper way to do this would be exception chaining,
"During handling of the above exception, another exception occurred",
but the current C API for that is complicated.)