Adds a pyinstaller hook for easier distribution
When moderngl is built into a pyinstaller executable with default settings, it fails at runtime with this error `ModuleNotFoundError: No module named 'glcontext'`

This is because Pyinstaller can only follow imports in Python, and glcontext is imported in C++ code.

To get around this, this commit adds a pyinstaller hook that adds glcontext as a hidden import, tipping pyinstaller off to add that to the build.

Users can achieve this same outcome with Pyinstaller's `--hidden-import` flag, but this will do it automatically so users don't need to worry about it.